Fixes based on public feedback
Here are many fixes

1. Remove UnityEngine from all my Systems and Components.
 - I created RMC.Common.UnityEngineReplacement.Vector3 and a few
others. This subject is a bit of a rabbit hole :) Probably such classes
already exist in the community for a larger project
 - Exception: I do call CoroutineUtility.StartCoroutine(), which is
dependent on UnityEngine, from within a few of my Systems. This is ok
for now :)
 - Status: **Done**
1. Update my use of 'ensureComponents'.
 - Removed all usages. Not needed (yet)
 - Status: **Done**
1. Address answers to my Q: "What is the best way to address OPTIONAL
components like Tick and Friction below?"
 - I simply added both Tick and Friction to the entity and remove the
if's from the Execute()
 - Status: **Done**
1. I like my approach to System Testing.
 - No changes needed.
 - Status: **Done**
1. "Features vs System?" I'll keep my code as is. I already recommended
it be added to the wiki on the wiki thread.
 - No changes needed.
 - Status: **Done**
1. Per Pausing, I see some alternatives in other projects including
GuardedSystems, but I like how I handle pausing.
 - No changes needed.
 - Status: **Done**
1. Remove attribute in DestroyComponent.cs and confirm class gets
generated without it.
 - I confirm it does generate without it. Then I added it back because
I prefer entity.WillDestroy(true).
 - Status: **Done**